requirements.txt
# Core Dependencies
biopython>=1.79
numpy>=1.21.0
pandas>=1.3.0
rdkit>=2022.03.1

# Visualization (optional, for analysis)
matplotlib>=3.4.0

# Web requests (for PDB API access)
requests>=2.26.0

# HTML parsing (for PDB web scraping if needed during cross docking set generation)
beautifulsoup4>=4.10.0

# Note: External tools required (install separately):
# - MMseqs2: https://github.com/soedinglab/MMseqs2
# - OpenBabel: http://openbabel.org/
